Animator component added. Scene can not be started and stopped.

+++ Insight/inc/Engine/Platform/Win32/Win32Platform.h
@@ -83,7 +83,7 @@ public:
     static void Assert(const char* message, const char* file, int line);
     static void CheckFailed(const char* message, const char* file, int line);
 
-    static void Sleep(i32 milliseconds);
+    static void Sleep(float milliseconds);
     static double GetTimeSeconds();
 
 	static U128 GenerateUUID();
